Refactor search functionality by restructuring classes and namespaces
